LearnSpere__IIC/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
groq==1.0.0
pyttsx3==2.90
Pillow==10.0.0
requests==2.31.0
werkzeug==2.3.7
beautifulsoup4==4.12.2
transformers>=4.30.0
torch>=2.0.0
scikit-learn>=1.2.0
numpy>=1.21.0
joblib>=1.2.0
sentencepiece  # Required for some Hugging Face models
PyJWT>=2.0.0
